requirements.txt
munch
slugify
pycrypto
termcolor
ruamel.yaml
python-magic
